api/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
drf_spectacular==0.29.0
